PLATE-26 Change listening port and file location
